# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API que gerencia os usuários, as reservas, os espaços e os pagamentos",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/bookings": {
            "get": {
                "description": "Retorna um JSON com todas as reservas",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Mostra todas as reservas",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.BookingParsed"
                            }
                        }
                    },
                    "404": {
                        "description": "Nenhuma reserva encontrada",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Reserva um espaço",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Adiciona uma nova reserva",
                "parameters": [
                    {
                        "description": "Dados da reserva",
                        "name": "booking",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bookings.BookSpaceRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "ID da reserva criada",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid booking or bad JSON",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Wrong password",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User or space not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Booking conflict",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/bookings/{id}": {
            "get": {
                "description": "Retorna um JSON com os dados da reserva",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Retorna a reserva do ID especificado",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.BookingParsed"
                        }
                    },
                    "404": {
                        "description": "Reserva não encontrada",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Cancela uma reserva",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Deleta a reserva especifcada pelo ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da reserva",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Credenciais do usuário",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bookings.AuthRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reserva cancelada",
                        "schema": {
                            "$ref": "#/definitions/models.BookingParsed"
                        }
                    },
                    "400": {
                        "description": "JSON inválido",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Senha incorreta ou usuário sem permissão",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Usuário ou reserva não encontrada",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Erro interno do Servidor",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/payments/{id}": {
            "get": {
                "description": "Retorna um JSON com os dados do pagamento referente ao ID informado",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Retorna a entidade de um pagamento específico",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do pagamento",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do pagamento encontrados",
                        "schema": {
                            "$ref": "#/definitions/models.Payment"
                        }
                    },
                    "404": {
                        "description": "Pagamento não encontrado",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Realiza o pagamento referente ao ID informado, usando o valor enviado no corpo da requisição",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Executa o pagamento de um ID específico",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do pagamento",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados do Pagamento",
                        "name": "payment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payments.PaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pagamento efetuado com sucesso",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "400": {
                        "description": "Pagamento inválido ou JSON mal formatado",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "bookings.AuthRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "bookings.BookSpaceRequest": {
            "type": "object",
            "properties": {
                "bookingTime": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "space": {
                    "type": "integer"
                },
                "startDate": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.BookingParsed": {
            "type": "object",
            "properties": {
                "days": {
                    "type": "integer"
                },
                "endDate": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "spaceId": {
                    "type": "integer"
                },
                "startDate": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "chave que aparece no JSON",
                    "type": "string"
                }
            }
        },
        "models.Payment": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "payedValue": {
                    "type": "integer"
                },
                "totalValue": {
                    "type": "integer"
                }
            }
        },
...
        "payments.PaymentRequest": {
            "type": "object",
            "properties": {
                "value": {
                    "type": "integer"
                }
            }
        },
...
    }
}