# server/internal/controller/payments/MakePayment.go
package payments

import (
	"encoding/json"
	"fmt"
	"net/http"
	"strconv"

	svc "github.com/d4vi13/SeuCantinho/server/internal/services/payments"
)

// MakePayment godoc
// @Summary Executa o pagamento de um ID específico
// @Description Realiza o pagamento referente ao ID informado, usando o valor enviado no corpo da requisição
// @Tags Payments
// @Accept json
// @Produce json
// @Param id path int true "ID do pagamento"
// @Param payment body PaymentRequest true "Dados do Pagamento"
// @Success 200 {object} models.ErrorResponse "Pagamento efetuado com sucesso"
// @Failure 400 {object} models.ErrorResponse "Pagamento inválido ou JSON mal formatado"
// @Failure 500 {object} models.ErrorResponse "Erro interno do servidor"
// @Router /payments/{id} [post]
func (controller *PaymentsController) MakePayment(w http.ResponseWriter, r *http.Request) {
	var req PaymentRequest

	id, err := strconv.Atoi(r.PathValue("id"))
	if (err != nil) || (id < 1) {
		http.NotFound(w, r)
		fmt.Printf("Failed to parsing req\n")
		return
	}

	err = json.NewDecoder(r.Body).Decode(&req)
	if err != nil {
		http.Error(w, "Invalid JSON", http.StatusBadRequest)
		return
	}
	defer r.Body.Close()

	ret := controller.paymentsService.MakePayment(id, req.Value)

	w.Header().Set("Content-Type", "application/json")
	switch ret {
	case svc.Success:
		w.WriteHeader(http.StatusCreated)
		json.NewEncoder(w).Encode(map[string]string{"info": "payment made"})
		fmt.Printf("INFO: payment %d was payed\n", id)
	case svc.InvalidPayment:
		w.WriteHeader(http.StatusBadRequest)
		json.NewEncoder(w).Encode(map[string]string{"error": "invalid payment attempted"})
		fmt.Printf("ERROR: invalid payment %d\n", id)
	default:
		w.WriteHeader(http.StatusInternalServerError)
		json.NewEncoder(w).Encode(map[string]string{"error": "unknown status"})
		fmt.Printf("ERROR: Unknown Status\n")

	}
}